requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
distro==1.9.0
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
openai==1.35.13
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.7
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
